tdf#134881 sw table dlg: make column change %-proportional
Apparently, there is a discrepancy in how the
various help manuals describe this.
What was actually happening was that the same
amount was being changed on each column
instead of the same percentage. That is really
bad if there are small columns, and a large one is reduced.
Plus it seemed like there were implementation problems anyway,
especially when growing a column too much.

This patch sets it to how the guide describes it.

"Adjust columns proportionally results in all columns changing
their widths by the same percentage when one is changed.
For example, if you reduce by half the size of a column,
the sizes of all the other columns will be halved."
(July 2018 LibreOffice Writer Guide 6.0 page 282)

Online help since 2013 simply says.
"Adjust columns proportionally:
If possible, change in column width will be equal for each column."
https://bugs.documentfoundation.org/show_bug.cgi?id=42023

One situation I noticed but don't think it needs dealing with
is if "from left" contains a value, then perhaps the table
should not be allowed to take up the full width?
It could grow to spill out past the edge of the page.
(Problem already existed with previous solution.)

Also, merged cells might cause some unexpected results,
but that becomes a user problem to fix I say.

I didn't think it would matter if I grew the table
a couple of Twips larger than full width. If it is,
then probably just reduce maxWidth by a twip for every
two columns - or truncate instead of round.
(And then I also need to account for theoretical columns
that are "increased" to MINLAY size too.)

Note that the table will NOT grow beyond
100% text width. However, if a table is already
larger than 100% (which the UI makes difficult to do)
at least an attempt to grow the columns will not cause
them to shrink. They will just stay the same sime.

make UITest_sw_table UITEST_TEST_NAME=\
tdf134881_colProportionalAdjust.tdf134881.test_tdf134881

Change-Id: Id57d80d9658b9ac1938189c75da0cbf309fc321a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/107604
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>