tdf#138546 sw PageDialog: fix bad reset of colwidth
The presence of border padding was causing the algorithm
to reset the overall width, which in page style triggered
an even redistribution of the columns. That should only
happen when the number of columns changes, or else one of
the 5 predefined column choices is selected.

My first python unit test is included.
make UITest_writer_tests4 \
UITEST_TEST_NAME=tdf138546.tdf138546.test_tdf138546

Change-Id: I650e429568f2ab794649cc557998a809479b6e82
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108210
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>