Make page navigation in move navigation handler consistent
Page navigation in the move navigation handler scrolls the view without
moving the cursor or unselecting selections. This patch makes page
navigation that results in document wrapping, scroll the view without
changing the cursor position or selections. This is useful, for example,
when find all search is made and page viewing is wanted that does not
destroy cursor selections.

Change-Id: Ie062eeba9a29c2b8a42605ed842d60244cc74147
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108380
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>