tdf#138544 sw LoadUserSettings: default SubtractFlysAnchoredAtFlys
This is the third patch in the series.
When you start a new document, Writer compat setting
"Tolerate white lines of PDF page backgrounds
for compatibility with old documents" is off.
Now, when saving and reloading it with LoadUserSettings disabled,
it is still the program default of "off", instead of "on".

Change-Id: I06869600f0d75c9d5457372d56e39750ff5b3dae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108547
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>