vcl: create PDF from DataContainer, use that when swapping in
VectorGraphicLoader is a set of functions that "load" a vector
graphic (starting with PDF) from a BinaryDataContainer. Usually
this only needs to create VectorGraphicData and set the correct
type, as loading will be done on-demand.

Use the VectorGraphicLoader function when swapping-in so this is
an alternative to calling LoadNative on GfxLink, which loads
into a new Graphic via. GraphicFilter, which is something we
really want to avoid.

Change-Id: If529c8ddca2b31660da4ea28d8e75bb20bf3c778
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109598
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>