tdf#137083 consider negative width in DrawPageViewGrid
SdrPageView::DrawPageViewGrid is called too for a ScDrawPage of a
RTL-sheet. In that case both the width of the SdrPage and the left
edge of the passed rectangle are negative. The x-values for drawing
the grid were wrong and a right-to-left sheet did not show the grid.
The patch adds a case distinction with correct x-values for case RTL.

Change-Id: I6d0c15bf7bbe8aff6ab2e72a440ba81f2e0e2281
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/108989
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>