tdf#139696 Use getToolboxId to set item id
This code depends on ToolboxController::m_nToolBoxId being already set, which
is the case for toolbars, as passed by ToolBarManager::CreateControllers via
the "Identifier" property, but not for the notebookbar, as can be seen in
sfx2::sidebar::ControllerFactory::CreateToolBarController.

Avoid this problem by using getToolboxId, like in most other toolbar controls.

Change-Id: I5dfb1bda0886ce3542274a422f8c9111e7fdcf7e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109722
Tested-by: Jenkins
Tested-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>