qtcreator: Drop some extra whitespace in .pro.user template
Fix indentation and drop some extra empty lines that are
automatically removed once the project is loaded in
Qt Creator and then closed anyway.

This will make it a little easier to compare what the initially
generated file looks like compared to the one as it is written
after (implicitly or explicitly) carrying out the configuration
in Qt Creator after loading the project.

Change-Id: I52173162efe635e6fb3e7f889c42ed67271c05b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110535
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>