devtools: remove mxObject from nodes and use maAny instead
First hunch was to always store an object (XInterface) in the
nodes, but as we can now handle values, it is much more convenient
to change that to an Any, which can store all the supported types.

This change refactores the code and removes mxObject from all
the object nodes. Instead getObjectFromAny is used in the places
where the mxObject was used before.

Change-Id: Ibf4a28d52102f99fefc1f7a7775b8084b739eb90
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110743
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>