tdf#118711 doc import: don't hardcode default page description
well, at least not at the beginning of the document.

This is a .DOC followup to the 7.1 .DOCX patch
in commit 8787a45f9cbb5dce61b20817ef0e549b5a227a95.

The problem with specifying is that it becomes a property
of the paragraph - and so if the paragraph is copied or moved,
the page break comes along with it and the user has to remove
the unnecessary page break wherever it lands.

Since the default will be applied anyway,
there is no value in specifying it.

In DOC, it appears that the first page is ALWAYS set
to the default page. I tried to create a title page/first page
page style, but it always round-tripped as default page.
That makes it easier, with only specified page number exception.

Change-Id: Ie2e5ec42331d66f1883fc7b96fcefceb6d32c73d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111240
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>