devtools: support XIndexAccess and XNameAccess in object inspector
If the current object in the object inspector is an container
that supports XNameAccess or XIndexAccess, we now show the named
or index values in the property view. This way the user can
navigate to sub-objects, which is sometimes needed because we get
an object supporting this intefaces, but can't navigate further.

Best example would be XShapeCollection, which is what you get when
selecting a shape (because with a selection we can select multiple
shapes), but it is not possible to navigate to the shape without
this change.

Change-Id: I6a18723eccbc41519e3eacf68bc5b6488e02fe22
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111529
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>