tdf#140572 docx export: avoid DocDefault superscript crash
This fixes a crash situation introduced in LO 6.4
by commit d71cf6390a89ea6a4fab724e3a7996f28ca33661.

The crash was occurring when the DocDefaults were
being exported. Since there was no character or
paragraph context, the lookup of the related
fontsize was crashing.

Moving this variable closer to where it is used
"just happens" to fix the crash. Because the
DocDefault is considered a style, we are just
exporting the superscript/subscript as a default
size, and so don't need to know the font size.
Thus a style-context position can't cause the
crash again.

So it isn't a great fix, and leaves the possibility
for the same thing to happen in other situations
(like RES_CHRATR_BACKGROUND), but at least it
quickly solves this known problem, and is easily
backportable, as well as following best practice
of keeping variables as close to where they are
used as possible - eliminating some unnecessary
lookups.

I did look for a more generic fix (or whether
other properties could be affected) and didn't
come with any clear examples or a better fix.

Change-Id: I03dc3d74a5148973c065cadb09c5c79068f12df1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111403
Tested-by: Jenkins
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>